% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.index.R
\name{compute.index}
\alias{compute.index}
\title{Compute index}
\usage{
compute.index(
  n,
  p,
  nval,
  methods,
  nrep = 20,
  seed = NULL,
  verbose = FALSE,
  file = NULL,
  rho.vec = c(0.35),
  s = 5,
  beta.type = 1,
  snr.vec = c(0.8, 1.6, 2.4, 3.6, 4.8)
)
}
\arguments{
\item{n}{The number of training observations}

\item{p}{The number of predictors.}

\item{nval}{The number of validation observations.}

\item{methods}{Methods to use.It could be "Lasso","Abess","Forward stepwise" or "Relaxed lasso".}

\item{nrep}{Number of repetitions of which to average the results. Default is 20.}

\item{verbose, file.rep, s, seed, beta.type}{Arguments to pass to \code{sim.master}}

\item{file}{Name of a file to save simulation results.}

\item{rho.vec}{Choose of rho.}

\item{snr.vec}{Choose of snr.}
}
\value{
A list with index(saves in file)
}
\description{
Compute index for simulation
}
\details{
Compute index.
}
