% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{Fmean}
\alias{Fmean}
\title{Frechet mean}
\usage{
Fmean(Y)
}
\arguments{
\item{Y}{A shapes data (landmark\*dim \*num)}
}
\value{
A value in shapes space.
}
\description{
Compute Frechet mean for shapes data
}
\examples{
\dontrun{
# Compute Frechet mean
data("brains")
Y=brains$x[,,1:10]
Fm=Fmean(Y)
Fm
}
}
