% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{gibbsR}
\alias{gibbsR}
\title{A Gibbs sampler using R}
\usage{
gibbsR(a, b, n, N)
}
\arguments{
\item{a}{param for original distribution}

\item{b}{param for original distribution}

\item{n}{param for original distribution}

\item{N}{the number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using R
}
\examples{
\dontrun{
rnR <- gibbsR(1,1,25,10000)
par(mfrow=c(2,1));
plot(rnR[,1],type='l')
plot(rnR[,2],type='l')
}
}
