% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_index}
\alias{plot_index}
\title{plot function}
\usage{
plot_index(
  file.list,
  method.nums = NULL,
  method.names = NULL,
  what = "error",
  main = NULL,
  ylim = NULL,
  legend.pos = "right",
  make.pdf = FALSE,
  fig.dir = ".",
  file.name = "sim",
  w = 8,
  h = 10
)
}
\arguments{
\item{method.nums}{The indices of the methods.}

\item{method.names}{The names of the methods.}

\item{what}{index for compare. It can be "error", "risk", "prop","F", or "nonzero".Default is "error".}

\item{legend.pos}{location of legend}

\item{make.pdf}{save as pdf}

\item{fig.dir, file.name}{The figure directory and file name.}

\item{w, h}{The width and height. Defaults are 8 and 10}

\item{file}{file list which contains simulation data}
}
\value{
figure or pdf
}
\description{
Plot the results for simulation
}
