% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{R2}
\alias{R2}
\title{R square}
\usage{
R2(Y, X)
}
\arguments{
\item{Y}{A shapes data (landmark\*dim \*num)}

\item{X}{variables in euclidean space}
}
\value{
A \code{list} object comprising:
\item{R2}{Determination coefficient.}
\item{y}{Adjust determination coefficient.}
}
\description{
Determination coefficient for model
}
\examples{
\dontrun{
# Compute determination coefficient
data("brains")
Y=brains$x[,,1:10]
X=cbind(brains$age,brains$grp)[1:10,]
r2=R2(Y,X)
r2
}
}
