% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{sim.data}
\alias{sim.data}
\title{Generate simulated data}
\usage{
sim.data(
  n,
  p,
  support.size = NULL,
  rho = 0,
  beta = NULL,
  family = "gaussian",
  snr = 10,
  sigma = NULL,
  seed = 1
)
}
\arguments{
\item{n}{The number of observations.}

\item{p}{The number of predictors.}

\item{support.size}{The number of nonzero coefficients.}

\item{rho}{The pairwise correlation in predictors. Default is \code{0}.}

\item{beta}{The coefficient values.}

\item{family}{The distribution of the simulated response.}

\item{snr}{Signal-to-noise ratio (SNR).Default is \code{snr = 10}.}

\item{sigma}{The variance of the gaussian noise. Default \code{sigma = NULL}}

\item{seed}{random seed. Default: \code{seed = 1}.}
}
\value{
A \code{list} object comprising:
\item{x}{Design matrix of predictors.}
\item{y}{Response variable.}
\item{beta}{The true coefficients.}
}
\description{
Generate simulated data under the linear model
}
\examples{
\dontrun{
# Generate simulated data
n <- 70
p <- 30
support.size <- 10
data<- sim.data(n, p, support.size)
}
}
