\name{StatComp21055-package}
\alias{StatComp21055-package}
\alias{StatComp21055}
\docType{package}
\title{
  The final review for 'StatComp' course
}
\description{
  This is the final review for 'StatComp' course, 
  including homework and some my own research function.
}

\author{
Jingguo Lan, ljg07158273@mail.ustc.edu.cn.

Maintainer: Jingguo Lan <ljg07158273@mail.ustc.edu.cn>
}

\keyword{ package }

